Refactor the point at which the memory cache is constructed
Addresses #3

The memory cache was constructed during the builder steps, which was
unintuitive and violated the builder expectation. It also caused the debug
flag to fail if it was specified after the memoryCache() step in the building
process.

By refactoring, we construct the memory cache only at the point the actual builder
constructs a Repo instance which also lets memoryCache() and debug() appear
anywhere in the builder steps.