Repo instances are no longer bound to a single type
A Repo instance can now cache any amount of arbitrary types. Type safety
is enforced and assumed correct based on the key used to access the
data. This trades ease of usage for some API flexibility, and slight
memory savings in the event that one used to be allocating multiple Repo
instances for different formats of data.

Even though there are pitfalls associated with backing Repos with
generic data structures, it should help encourage the usage of a single
Repo instance per service instead of duplicating multiple Repos across a
single service - which was needlessly increasing memory usage and
program overhead.