Completely restructure how Repo instances are handled
A Repo by default will track against a single upstream.
This does away with the key based uniqueness of requests and will
guarantee that upstream data is only interacted with inside of a single
flow.

Multiple pieces of data can be cached using a new MultiRepo interface
which is a simple Map of Repo's which are keyed with a String for
uniqueness.

For migration, any repo that uses a 'throwaway' key of some kind
like a simple string can migrate to the cleaner Repo interface,
while any repo that used an actual unique key identified can
migrate to a MultiRepo interface