Create a simple but naive persister implementation
The persister implementation deals with files and will save
and restore lists of stringified data to the provided file.

The persister does not do any caching at its layer as it would
only be used if the memory cache fails to provide cached data,
so each read and write will directly interact with disk.

The persister can potentially be slow because of this.

The persister will not block on writes, meaning that reading
immediately after a write is not guaranteed to return the most
up to date data - a MemoryCache should be used for these smaller
operations