Remove the Repo manipulation methods
Removes push() pushAll() replace() and replaceAll()

These are generally anti-patterns. A Repo object should be
used with a single source of truth, and cache it's values

A Repo which is modified on the fly using one of the above
methods has now become a second source of truth which can
be different from its upstream.

As such, these methods are removed in the next release